RecordSystem, ScrenLimits, Shooter, damageManage
Record System Added, now you can improve yourself.

ScreenLimits Added, now the player cant go out of the screen.

Hold to Shoot Added, now you can hold and not onlyu press.

When you get damage, you are invulnerable to more damage during the blink animation.

Extra Changes:
- The Asteroid spawners spawn asteroids with probability now.

- New icon app build

+++ Assets/Scripts/DamageBehaviour.cs
@@ -1,10 +1,9 @@
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.Serialization;
 
 public class DamageBehaviour : MonoBehaviour
 {
-    [SerializeField] private float damage;
+    [SerializeField] private int damage;
     
     [SerializeField] private List<string> layersToDamage;
